% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ceda_data_path}
\alias{ceda_data_path}
\title{Title}
\usage{
ceda_data_path(
  ceda_data_folder = sample_ceda_data_folder(),
  country,
  metric = c("tmp", "tmn", "tmx"),
  version = 2020
)
}
\arguments{
\item{version}{}
}
\value{

}
\description{
Title
}
