% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_functions.R
\name{create_agg_cru_cy_df}
\alias{create_agg_cru_cy_df}
\title{Create a dataframe containing multiple cru_cy metrics}
\usage{
create_agg_cru_cy_df(agg_cru_cy_folder, agg_cru_cy_metrics, agg_cru_cy_year)
}
\arguments{
\item{agg_cru_cy_folder}{path to a folder contains all cru_cy data folders,
passed to the \code{cru_cy_folder} argument in \code{read_cru_cy_files}}

\item{agg_cru_cy_metrics}{a character string of metrics which is iterated over with \code{read_cru_cy_files}
using \code{lapply}}

\item{agg_cru_cy_year}{identifies the year of the CEDA data to loaded,
passed to the \code{cru_cy_year} argument in \code{read_cru_cy_files}}
}
\value{
a tibble with five columns: Country, Metric, Year, Month, Value.
In contrast to \code{read_cru_cy_files} there may be multiple metrics if \code{length(agg_cru_cy_metrics) > 1}.
}
\description{
Reads all cru_cy (Climate Research Unit Country) files for a multiple metrics and binds to create a single tibble.
Data is drawn from a directory supplied by the user, likely created through the down_cru_cy_files function.
}
