% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cru_cy_colnames}
\alias{cru_cy_colnames}
\title{cru_cy column names}
\format{
A string list with 18 entries.
\describe{
\item{YEAR}{The string "YEAR", representing the year represnted in the data.}
\item{JAN}{The string "JAN", representing a time period consisting of the month January.}
\item{FEB}{The string "FEB", representing a time period consisting of the month February.}
\item{MAR}{The string "MAR", representing a time period consisting of the month March.}
\item{APR}{The string "APR", representing a time period consisting of the month April.}
\item{MAY}{The string "MAY", representing a time period consisting of the month May.}
\item{JUN}{The string "JUN", representing a time period consisting of the month June.}
\item{JUL}{The string "JUL", representing a time period consisting of the month July.}
\item{AUG}{The string "AUG", representing a time period consisting of the month August.}
\item{SEP}{The string "SEP", representing a time period consisting of the month September.}
\item{OCT}{The string "OCT", representing a time period consisting of the month October.}
\item{NOV}{The string "NOV", representing a time period consisting of the month November.}
\item{DEC}{The string "DEC", representing a time period consisting of the month December.}
\item{MAM}{The string "MAM", representing a time period consisting of the months of March, April, and May.}
\item{JJA}{The string "JJA", representing a time period consisting of the months of June, July, and August}
\item{SON}{The string "SON", representing a time period consisting of the months of September, October, and November.}
\item{DJF}{The string "DJF", representing a time period consisting of the months of December, Janaury, and February.}
\item{ANN}{The string "ANN", representing a time period consisting of the full year.}
}
}
\usage{
cru_cy_colnames
}
\description{
A string list containing column names of the raw CEDA cru_cy data.
}
\examples{
cru_cy_colnames
}
\keyword{datasets}
