% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down_functions.R
\name{down_cru_cy_data}
\alias{down_cru_cy_data}
\title{The function cru_cy_metrics downloads the latest (2020) Climate Research Unit
Country-Level datasets (cru_cy) (http://data.ceda.ac.uk/badc/cru/data/cru_cy/cru_cy_4.04/data)
from CEDA via FTP (ftp://ftp.ceda.ac.uk/badc/cru/data/cru_cy/cru_cy_4.04/data/).
Users must first create an account with CEDA.
The following metrics are available, and are denoted by three letter codes:
"cld" = Cloud cover,
"dtr" = Diurnal temperature range,
"frs" = Ground frost frequency,
"pet" = Potential evapotranspiration,
"pre" = Precipitation,
"tmn" = Near surface temperature minimum,
"tmp" = Near surface temperature,
"tmx" = Near surface temperature maximum,
"vap" = Vapour pressure,
"wet" = Wet day frequency}
\usage{
down_cru_cy_data(username, password, cru_cy_metric, cru_cy_folder)
}
\arguments{
\item{username}{the username for a user-created CEDA account}

\item{password}{the password for a user-created CEDA account}

\item{cru_cy_metric}{the three letter metric code the user wishes to download}

\item{cru_cy_folder}{the filepath for the data to be downloaded in}
}
\value{

}
\description{
The function cru_cy_metrics downloads the latest (2020) Climate Research Unit
Country-Level datasets (cru_cy) (http://data.ceda.ac.uk/badc/cru/data/cru_cy/cru_cy_4.04/data)
from CEDA via FTP (ftp://ftp.ceda.ac.uk/badc/cru/data/cru_cy/cru_cy_4.04/data/).
Users must first create an account with CEDA.
The following metrics are available, and are denoted by three letter codes:
"cld" = Cloud cover,
"dtr" = Diurnal temperature range,
"frs" = Ground frost frequency,
"pet" = Potential evapotranspiration,
"pre" = Precipitation,
"tmn" = Near surface temperature minimum,
"tmp" = Near surface temperature,
"tmx" = Near surface temperature maximum,
"vap" = Vapour pressure,
"wet" = Wet day frequency
}
\examples{

}
