% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_functions.R
\name{read_cru_cy_file}
\alias{read_cru_cy_file}
\title{Read a single CEDA cru_cy text file}
\usage{
read_cru_cy_file(cru_cy_file)
}
\arguments{
\item{cru_cy_file}{a file path to a .per text file downloaded from CEDA}
}
\value{
a tibble containing climate data for a single country
}
\description{
Establishes a function which reads an individual CEDA .per (text) file
}
