% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_functions.R
\name{read_cru_cy_files}
\alias{read_cru_cy_files}
\title{Read all CEDA file for a particular metric and bind create a single tibble}
\usage{
read_cru_cy_files(cru_cy_folder, cru_cy_metric, cru_cy_year)
}
\arguments{
\item{cru_cy_folder}{path to a folder contains cru_cy data files.}

\item{cru_cy_metric}{three-letter code used to identify filename to read and write.}

\item{cru_cy_year}{identifies the year of the CEDA data to loaded.}
}
\value{
a tibble with columns
}
\description{
Zeke will add more description here.
}
