% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sample_ceda_data_folder}
\alias{sample_ceda_data_folder}
\title{Establish a user-independent file path to the sample CEDA cru_cy data bundled in CEDATools}
\usage{
sample_ceda_data_folder()
}
\value{
a file path to bundled CEDA cru_cy data for tmp (Near surface temperature),
tmn (Near surface temperature minimum), tmx (Near surface temperature maximum)
for Ghana and South Africa.
This path is unique indpendent of the user/machine.
}
\description{
Establish a user-independent file path to the sample CEDA cru_cy data bundled in CEDATools
}
