% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sample_ceda_data_path}
\alias{sample_ceda_data_path}
\title{Title}
\usage{
sample_ceda_data_path(
  country = c("Ghana", "South_Africa"),
  metric = c("tmp", "tmn", "tmx"),
  version = 2020
)
}
\arguments{
\item{metric}{One of "tmp" (for mean monthly temperature),
"tmn" (for mean monthly daily minimum temperature), or
"tmx" (for mean monthly daily maximum temperature).}

\item{version}{An integer representing the year in which CEDA data were released.
E.g., \code{2020} means data up through year 2019 that were released in the year 2020.}
}
\value{

}
\description{
Title
}
