% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ceda_data_path}
\alias{ceda_data_path}
\title{Create a list of file paths to CEDA cru_cy files}
\usage{
ceda_data_path(
  ceda_data_folder = sample_ceda_data_folder(),
  country,
  metric = c("cld", "dtr", "frs", "pet", "pre", "tmn", "tmp", "tmx", "vap", "wet"),
  version = 2020
)
}
\arguments{
\item{ceda_data_folder}{The file path to the folder containing the cru_cy CEDA Data,
specificially a folder named "CEDA Data".
By default this is set to \code{sample_ceda_data_folder},
but any file path leading to the "CEDA Data" folder can be used.
Note that the data must be present in a folder in CEDA Data/CEDA_2020
with the corresponding metric code.}

\item{country}{Any of the countries present in cru_cy CEDA files,
note that some country names are unique to CEDA,
as they are taken from the file names, and contain underscores,
and abbreviations such as "Isl" for "Islands".}

\item{metric}{One of:"cld" = Cloud cover,
"dtr" = Diurnal temperature range,
"frs" = Ground frost frequency,
"pet" = Potential evapotranspiration,
"pre" = Precipitation,
"tmn" = Near surface temperature minimum,
"tmp" = Near surface temperature,
"tmx" = Near surface temperature maximum,
"vap" = Vapour pressure,
"wet" = Wet day frequency}

\item{version}{An integer representing the year in which CEDA data were released.
E.g., \code{2020} means data up through year 2019 that were released in the year 2020.}
}
\value{
a list of file paths for selected cru_cy CEDA text files (.per), determined by metric and country.
}
\description{
Create a list of file paths to CEDA cru_cy files
}
